etl_pipeline/requirements.txt
dagster==1.3.13
dagit==1.3.13
dagster-postgres==0.19.13 
pendulum==2.1.2
minio==7.2.7
pyarrow
cryptography==42.0.5
psycopg2-binary==2.9.9
dbt-core
dbt-postgres
pymysql
pandas==1.5.3
SQLAlchemy==1.4.46
numpy==1.26.3
pyspark==3.3.2
polars